GEODE-8302: Fixed 'events not queued conflated' stats when group-tran… (#5313)
This is a fix on top of GEODE-7971 for a problem with serial gateway senders
using the group-transaction-events which could lead in very unlikely situations
to not replicate some events.

* GEODE-8302: Fixed 'events not queued conflated' stats when group-transaction-events is true

Batch conflation is not compatible with group-transaction-events.
It must be prevented that both are enabled at the same time for
a given gateway sender.

* GEODE-8302: Add extra checks in propagation with HA tests and add fix in conflation stats

* GEODE-8302: Update GatewaySenderFactory javadoc

Added information about incompatibility between
group-transaction-events and enable-batch-conflation.

* GEODE-8302: Doc changes after review

* GEODE-8302: Fix for failing testReplicatedSerialPropagationHAWithGroupTransactionEvents

* GEODE-8302: doc and error messages changes after review

* GEODE-8302: Fix failing test case and possibly fix GEODE-8320