Fix null pointer exception with group transaction events
The null pointer exception can appear when
adding missing events to the batch if group-transaction-events
is enabled and there is a mix of puts with transactions and
without transactions.
When this happens, the batch is finally sent but the
second time the batch is sent ignoring the group-transaction-events
flag. There could be other secondary effects.