GEODE-8856: Persist gateway-sender state
* Persist gateway-sender state to cluster configuration

* Impacted commands:
    - stop gateway-sender   (state = stopped)
    - start gateway-sender  (state = running)
    - pause gateway-sender  (state = paused)
    - resume gateway-sender (state = running)

* Read state parameter from cluster configuration and bring
  gateway-sender to that state after member started up

* Introuduce new distributed tests for all commands

Remove uneccessary code

Fix for distributed and integration tests

Fix for flaky test

Remove unused code

Added new distributed test

* new test for parallel gateway-sender

Distributed test updated

Added tests for serial and parallel gateway sender. These tests verify
that gateway-sender stat is persisted in cluster configuration, and also
that gateway-sender works as expected after the restart of member.

Added small fix and unit tests:

 * Minor fix in the ResumeGatewaySenderCommand.java
 * CacheXmlParserTest.java
 * PauseGatewaySenderCommandTest.java
 * ResumeGatewaySenderCommandTest.java
 * StartGatewaySenderCommandTest.java
 * StopGatewaySenderCommandTest.java

Fix for following case:

 1. Create gateway-sender on all servers with manual-start set to false
 2. Pause gateway-sender using "pause gateway-sender" command.
 3. Stop and then start again all servers and verify that gateway-sender remain paused.
 4. Resume gateway-sender and verify it's state
 5. Stop gateway-sender and verify it's state
 6. Start gateway-sender and verify it's state

The problem was that in step 6. the gateway-sender was started in paused
state, because flag "startEventProcessorInPausedState" was set in step 3.
and it was never cleared afterwards. This fix clears this flag after
gateway-sender is started in paused state at step 3.

Removed PowerMock from unit test

Recover paralell gateway-sender queue in stopped state

This commit implements scenario when paralell gateway-sender
is recovered in stopped state after the restart of server. Basically,
the gateway-sender is brought up to the same state as it would be
after it is started and then stopped using gfsh commands. This
way gateway-sender queue region doesn't block the collocated data
region to reach online status.

Additionally, new dunit test are introduced to verify above behavior:
- ParallelGatewaySenderOperationClusterConfigDUnitTest
- SerialGatewaySenderOperationClusterConfigDUnitTest

Fix for failing test cases:

- fix for testThatPauseStateRemainAfterTheRestartOfMembersRedundantPersistent

   The batch dispatcher thread blocks and waits for configured time (batch-time-interval) to
   read new events. The batch-time-interval default value is 1000 milliseconds. So if
   gateway-sender is paused and traffic is received during that period of 1000 milliseconds,
   the dispacher thread will collect those event in batch and dispatch them
   to remote site. After 1000 milliseconds expire dispatcher thread will be
   actually paused. So it is necessary to wait for batch-time-interval + offset,
   before putting new data.

- other tc's

  recoverInStoppedState() is rejected if locators aren't started. This
  check is now removed. It is not necessary to have available locators
  when gateway-senders is recovered in stopped state.

Update after comments and fix for TC's

comments:
- "collocated" replaced with "colocated"
- InternalGemFireException is thrown in case of invalid state value

TC's:
- "test_ValidateParallelGatewaySenderAttributes" was missing a disk store.
   When manual-start=true, server will create persistent partition region,
   so it is neccesary to first create disk-store.

- "testThatPauseStateRemainAfterTheRestartOfMembersRedundantPersistent"
  The batch dispatcher thread blocks and waits for configured time (batch-time-interval) to
  read new events. The batch-time-interval default value is 1000 milliseconds. So if
  gateway-sender is paused and traffic is received during that period of 1000 milliseconds,
  the dispacher thread will collect those event in a batch and dispatch them
  to remote site. After 1000 milliseconds expire dispatcher thread will be
  actually paused. So it is neccessary in this TC to wait for all dispatcher
  threads to be actually paused before sending new traffic.

Fix for LGTM

- Also to restart Distributed and Acceptance test

Minor improvment and fix for TC's that result with timeout exceeded

Improves solution related to recovery in stopped state. Parallel gateway-sender
queue will be created/recovered at gateway-sender creation only if colocated data
region is also available. If data region is not available, then parallel gateway
sender queue will be created/recovered after colocated data region is created/recovered.

Fixed TC: testpersistentWanGateway_restartSender_expectAllEventsReceived_scenario2

This TC was failing because gateway-sender queue persistent partition region
is destroyed on all members when gateway-sender is started with clean queues,
and due to that events were not replicated to the other site. This happened because
now gateway-sender partition persistent queue is recovered/created when manual-start
set to true during server startup. The reason why this worked before this improvement
is  because partition region was not automatically created/recovered during server startup,
but only after gateway-sender is started with start command. So in that case when
"start gateway-sender with clean queue" is executed, the queue region is not destroyed,
since it doesn't exist, and therefore queued data remains on the other members.
I have changed this case to reflect the new behavior, and to check that everything works
as expected after gateway-sender is started with clean queues.

other TC's:

The other TC's were failing because partitioned persistent regions were created
in synchronous way. Now partition persistent queue region is created/recovered
during the creation of data region, so it is now necessary that partition region
is created in asynchronous way, so that members don't infinitely wait for member
with most up to date persistent data.

Fix for stress test